// client/app/child/child.ts
export class Child {
    name: {
        first: string;
        last: string;
    };
    password: string;
    parent: string;
}


// name: String,
// password: String,
// parent: String,
// activeQuests: [String],
// hero: {
//     name: String,
//     inventory: [String],
//     credits: Number,
//     xp: Number,
//     strength: Number,
//     wisdom: Number,
//     kindness: Number,
//     courage: Number,
//     responsibility: Number
// }